+++ rs/ledger_suite/common/ledger_canister_core/BUILD.bazel
@@ -21,6 +21,7 @@ rust_library(
         "//rs/types/management_canister_types",
         "//rs/utils",
         "@crate_index//:candid",
+        "@crate_index//:ic-cdk",
         "@crate_index//:serde",
     ],
 )
