+++ rs/ledger_suite/common/ledger_canister_core/src/runtime.rs
@@ -1,6 +1,6 @@
 use async_trait::async_trait;
 use candid::utils::{ArgumentDecoder, ArgumentEncoder};
-use ic_base_types::CanisterId;
+use ic_base_types::{CanisterId, PrincipalId};
 
 #[async_trait]
 pub trait Runtime {
@@ -32,4 +32,34 @@ pub fn heap_memory_size_bytes() -> usize {
 #[cfg(not(any(target_arch = "wasm32")))]
 pub fn heap_memory_size_bytes() -> usize {
     0
+}
+
+#[derive(Debug)]
+pub struct CdkRuntime;
+
+#[async_trait]
+impl Runtime for CdkRuntime {
+    fn id() -> CanisterId {
+        CanisterId::try_from(PrincipalId::from(ic_cdk::api::id())).unwrap()
+    }
+
+    fn print(msg: impl AsRef<str>) {
+        ic_cdk::api::print(msg)
+    }
+
+    async fn call<In, Out>(
+        id: CanisterId,
+        method: &str,
+        cycles: u64,
+        args: In,
+    ) -> Result<Out, (i32, String)>
+    where
+        In: ArgumentEncoder + Send,
+        Out: for<'a> ArgumentDecoder<'a>,
+    {
+        let principal_id = PrincipalId::from(id);
+        ic_cdk::api::call::call_with_payment(principal_id.into(), method, args, cycles)
+            .await
+            .map_err(|(code, msg)| (code as i32, msg))
+    }
 }
