+++ rs/ledger_suite/icp/ledger/BUILD.bazel
@@ -8,7 +8,6 @@ package(default_visibility = ["//visibility:public"])
     rust_library(
         name = "ledger" + name_suffix,
         srcs = [
-            "src/dfn_runtime.rs",
             "src/lib.rs",
             "src/tests.rs",
         ],
@@ -35,8 +34,6 @@ package(default_visibility = ["//visibility:public"])
             "//rs/ledger_suite/icp:icp_ledger",
             "//rs/ledger_suite/icrc1",
             "//rs/limits",
-            "//rs/rust_canisters/dfn_candid",
-            "//rs/rust_canisters/dfn_core",
             "//rs/types/base_types",
             "@crate_index//:candid",
             "@crate_index//:hex",
